feat: live history updates + speed/burst diagnostics
- History tab now re-reads JSONL every 0.5s while session is active
- Sub-second timestamp precision (ts + ts_nanos from SQLite)
- New timing heuristics:
  - effective TPS (tokens/sec wall-clock) -> FAST flag if >150
  - deltas/sec -> BURST flag if >50 δ/s
  - burst detection: 3+ deltas within 50ms window -> BURSTS flag
  - avg delta size (chars) -> CHUNK flag if >100
- Diagnostics table now shows TPS and δ/s columns
- Updated legend with speed metric descriptions

+++ src/reveal/models.py
@@ -35,7 +35,7 @@ class SessionGroup:
 class SSEEvent:
     """A single SSE event from the logs database."""
     id: int
-    timestamp: int
+    timestamp: float  # seconds with nanosecond precision
     level: str
     event_type: str
     raw_body: str
